git-protocol: rename crate to `link-git-protocol`
The mouthfulness of `radicle-link-git-protocol` was actually the
motivation for dropping the `radicle-` prefix in #763 and #760, so let's
do this now.
