+++ Cargo.toml
@@ -7,11 +7,11 @@ members = [
   "e2e",
   "git-ext",
   "git-helpers",
-  "git-protocol",
   "git-trailers",
   "librad",
   "link-canonical",
   "link-crypto",
+  "link-git-protocol",
   "link-identities",
   "macros",
   "seed",
