+++ link-git-protocol/src/packwriter.rs
