+++ link-git-protocol/src/take.rs
