+++ test/src/test/unit.rs
@@ -4,6 +4,6 @@
 // Linking Exception. For full terms see the included LICENSE file.
 
 mod git_ext;
-mod git_protocol;
 mod git_trailers;
 mod librad;
+mod link_git_protocol;
